+++ CHANGELOG.md
@@ -13,6 +13,13 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## [v2.1.0] - 2018-12-01
+
+### Added
+
+- Public method `\Jawira\PhingVisualizer\Diagram::transformToPuml`, this method
+is supposed to be used by other libraries or projects.
+
 ## [v2.0.0] - 2018-12-01
 
 ### Changes
@@ -178,3 +185,4 @@ executable usage
 [v1.0.0]: https://github.com/jawira/phing-visualizer/compare/v0.0.0...v1.0.0
 [v1.6.0]: https://github.com/jawira/phing-visualizer/compare/v1.5.1...v1.6.0
 [v2.0.0]: https://github.com/jawira/phing-visualizer/compare/v1.7.2...v2.0.0
+[v2.1.0]: https://github.com/jawira/phing-visualizer/compare/v2.0.0...v2.1.0